Fix some problems with run and index, spaghet works
There were a few issues with how repeated runs were handled, which seem
to be fixed now. There was also an issue where images ending on a run
did not have the last run output correctly since it was counting wrong.

The accelerator does not handle the end of an image gracefully though,
and will continue to try and output garbage if the cpu wants to keep
reading. This will need to be fixed.