Add WRITE_CPU state
This state gives a signal to the CPU that it can read back processed
data from the accelerator (The accelerator wants to "write" to main
memory).

There is an issue with this implementation, where the program memory and
images (main memory) are treated more like registers and the sram built
into the accelerator is more like, well, sram. In order to get readback
working, those memory areas will probably have to be changed to act more
like ROM/RAM and not like registers.