Fix bug with stale status value race condition
It was reading status to make sure that final was not set, then reading
status again to see if w_flag was set.

If it checked and saw that final flag was not set, then final flag (and
w_flag at the same time) became high, then it would see w_flag and try
to read an entire chunk.

Race condition is removed as status is only read once.